ExprDiff

Description

differentiable expressions Copyright : (c) Taha Khan 2018 License : WTFPL Maintainer : khant21mcmaster.ca Stability : experimental Portability : POSIX This module depends on the ExprType Module. It uses the Expr datatype and performs various functions over the differential expressions them. -- | The base functions include: -- * eval -- * simplify -- * partDiff -- | Additional functions include: -- * newton

Type Class Declaration

class DiffExpr a where

Class DiffExpr: - Differentiable Expressions - ---------------------------------------- - This class has methods over the Expr datatype - that assist with construction and evaluation of - differentiable expressions - ----------------------------------------- - Methods: - eval : takes a dictionary of variable identifiers and values, and uses it to compute the Expr fully EX. eval (Map.fromList [("x", 0)]) (Mult (Const 2) (Mult (Const 2) (Const 2))) => 8.0 - simplify : Takes a possibly incomplete dictionary and uses it to reduce Expr as much as possible EX. Add (Add (Var "x") (Const 1)) (Add (Const 2) (Var y)) => Add (Constant) (Add (Var "x") (Var "y")) <-- IDEALLY (Practically, this may not be the case) - partDiff : Given a var identifier, differentiate IN TERMS of that identifier EX. partDiff "x" (Mult (Neg (Const 2)) (Var "x")) => Add (Mult (Mult (Neg (Const 1.0)) (Const 0.0)) (Var "x")) (Mult (Neg (Const 2.0)) (Const 1.0)) - newton : Takes a dictionary of variable identifiers and values, a string which represents what variable to differentiate in terms of, an Expr a, and an a which respresents the intial guess for the x value of the root. EX. newton (Map.fromList [("x", 1)]) "x" (Add (Var "x") (Neg (Const 5))) 1 => 5.0 => Note: newton only performs one iteration of newtons method to approximate the root - Default Methods - !+, !*, var, val : are function wrappers for Expr constructors that - performs additional simplifications

Minimal complete definition

eval, simplify, partDiff, newton

Methods

eval :: Map String a -> Expr a -> a

Evaluate an expression given a dictionary of variables values

simplify :: Map String a -> Expr a -> Expr a

Simplify an expression by substituting in the variable values

partDiff :: String -> Expr a -> Expr a

Partially differentiate a fucntion with respect to a certain identifier

newton :: Map String a -> String -> Expr a -> a -> a

Approximate the roots of an expression using Newtons Method

(!+) :: Expr a -> Expr a -> Expr a

The Following are are default methods for all Expr types -

(!*) :: Expr a -> Expr a -> Expr a

(!/) :: Expr a -> Expr a -> Expr a

(!^) :: Expr a -> Expr a -> Expr a

expo :: Expr a -> Expr a

val :: a -> Expr a

var :: String -> Expr a

cos1 :: Expr a -> Expr a

sin1 :: Expr a -> Expr a

ln1 :: Expr a -> Expr a

Instances of the DiffExpr type class