ExprParser

Description

Documentation

parseExprDouble

Arguments

:: String
The string to be parsed

-> Expr Double
The resulting parsed expression

parses an expression into Expr Double. The syntax for parser are as follows. + to add, - to subtract, / to divide, * to multiply, ^ for exponentiation, cos for cosine, sin for sin, log for natural logarithm, e^ for natural exponential, any combination of alphabetical characters for a variable (expect those previously listed), any real numbers for constants, {} to indicate a matrix, , to seperate values in a row of a matrix, ; to seperate rows of a matrix, () to specify order of operations. Some Examples of valid inputs : "5+2^(2-5)" "5.223/(x*2)" "e^23" "log(dsafa/2)" "cos(sin(5.0-x))" "5*{2,3,4;1,2,3;2,3,4}"