ExprPretty

Description

ExprPretty uses parenns to return values with brackets, instance show is used to display what to return in accordance to the expression eveluated

Documentation

parenns :: String -> String

parenns takes a string and returns a string, the aim is to evaluate the epression with the brackets as an output