ExprType

Description

Documentation

data Expr a

This contains a type class for differentiable expressions and also type class for matrices.

Constructors

Add (Expr a) (Expr a)
This constructor represents the addition of two expressions of type 'Expr a'

Mult (Expr a) (Expr a)
This constructor represents the Multiplication of two expressions of type 'Expr a'

Sub (Expr a) (Expr a)
This constructor represents the Subtration of two expressions of type 'Expr a'

Div (Expr a) (Expr a)
This constructor represents the Divition of two expressions of type 'Expr a'

Const a
This constructor represents a Constant

Var String
This constructor represents a Variable where the Variable is a String

Cos (Expr a)
This constructor represents the cosine of an expression of type 'Expr a'

Sin (Expr a)
This constructor represents the sin of an expression of type 'Expr a'

Log (Expr a)
This constructor represents the sin of an expression of type 'Expr a'

Exp (Expr a) (Expr a)
This constructor represents the exponent of two expressions of type 'Expr a'

NatExp (Expr a)
This constructor represents the natural exponent (e^(Expr a) ) of an expression of type 'Expr a'

Matrix [[a]]
This constructor represents a matrix with a List of Lists. Where each List cotains the elements of that row. e.g Matrix [[1,2],[3,4]] is the representaion of the matrix 1 2 3 4

getVars :: Expr a -> [String]

getVars is a function that takes an expression and sees which variables (Var a) are present and returns a list of all the variables