SafeMath

Description

Documentation

class SafeMath a where

Minimal complete definition

myCos, mySin, myLog, myExp, myPow

Methods

myCos :: a -> a

Adapt the prelude cos function to all number types

mySin :: a -> a

Adapt the prelude sin function to all number types

myLog :: a -> a

Adapt the prelude log function to all number types

myExp :: a -> a

Adapt the prelude exp function to all number types

myPow :: a -> a -> a

Adapt the prelude (**) function to all number types