ExprDiff

Description

Documentation

eval

Arguments

:: DiffExpr a

=> Map String a
Takes in a variable dictionary

-> Expr a
Takes in an expression

-> a
Return its value, if possible

Get the value of expression. Return error if not enough information is given.

simplify

Arguments

:: DiffExpr a

=> Map String a
Takes in a variable dictionary

-> Expr a
Takes in an expression

-> Expr a
Return the simplified form of expression

Simplify the expression to minimal form.

partDiff

Arguments

:: DiffExpr a

=> String
Takes in the variable to be partially derived for, as a string

-> Expr a
Takes in the expression to be derived from

-> Expr a
Return the partially derived form of the expression

Perform partial differentiation to an expression with a given variable name.

diff

Arguments

:: DiffExpr a

=> Expr a
Takes in the expression you want to derive for

-> Expr a
Takes in the expression you want to derive from

-> Expr a
Return the derivative of the second expression for the first expression

Was designed as an auxiliary function for partDiff, but it can actually perform differentiation based on ANY expression, not just a single variable. The partDiff function is actually a special application of diff.

simplifiable

Arguments

:: DiffExpr a

=> Map String a
Takes in a variable dictionary

-> Expr a
Takes in an expression

-> Bool
Return True if it can be simplified. Return False if it cannot.

A function used to determine if an expression can be simplified or not, returns a Bool.

simplifyAdd

Arguments

:: DiffExpr a

=> Map String a
Takes in a variable dictionary

-> Expr a
Takes in an expression. Must be in Add form.

-> Expr a
Apply special tricks for simplifying Add types.

Auxiliary function used to simplify specifically Add types, and perform thorough simplification based on commutative laws.

simplifyMult

Arguments

:: DiffExpr a

=> Map String a
Takes in a variable dictionary

-> Expr a
Takes in an expression. Must be in Mult form.

-> Expr a
Apply special tricks for simplifying Mult types.

Auxiliary function used to simplify specifically Mult types, and perform thorough simplification based on commutative laws and associative laws.

simplifyPow

Arguments

:: DiffExpr a

=> Map String a
Takes in a variable dictionary

-> Expr a
Takes in an expression. Must be in Pow form.

-> Expr a
Apply special tricks for simplifying Pow types.

Auxiliary function used to simplify specifically Pow types.