ExprDiff

Description

Class Declaration For DiffExpr

class DiffExpr a where

Minimal complete definition

eval, simplify, partDiff, antiDeriv

Methods

eval :: Map String a -> Expr a -> a

Function that is used to evalute an expression given a dictionary

simplify :: Map String a -> Expr a -> Expr a

Function that is used to simplify an expression given a dictionary

partDiff :: String -> Expr a -> Expr a

Function that is used to calculate the partial derivative of an expression

antiDeriv :: Expr a -> Expr a

Function that is used to calculate the anti d of an expression

(!+) :: Expr a -> Expr a -> Expr a

These are the function that are used to apply operators such as + and - over Expr types

(!*) :: Expr a -> Expr a -> Expr a

(!/) :: Expr a -> Expr a -> Expr a

(!^) :: Expr a -> Expr a -> Expr a

val :: a -> Expr a

var :: String -> Expr a

newSin :: Expr a -> Expr a

newCos :: Expr a -> Expr a

newLn :: Expr a -> Expr a

newExp :: Expr a -> Expr a