ExprDiff

Description

differentiable expressions Copyright : (c) Kexin Liu 2018 License : WTFPL Maintainer : liuk44mcmaster.ca Stability : experimental Portability : POSIX

Documentation

class DiffExpr a where

Class DiffExpr:

Differentiable Expressions

----------------------------------------------

This class has methods over the Expr datatype

that assists with construction and evaluation

of differentiable expressions

--------------------------------------------------------

Methods:

eval - takes a dictionary of variable identifiers

and values, and uses it to compute the

Expr fully

simplify - takes a possibly incomplete dictionary

and uses it to reduce Expr as much as

possible

Add (Add (Var "x") (Const 1)) (Add (Const 2) (Var "y"))

=> Add (Const 3) (Add (Var "x") (Add (Var "y"))

eg. e1 = x + y

e2 = y + x

simplify e1 == simplify e2

partDiff - given an var identifier, differentiate

IN TERMS of that identifier

Default Methods - !+, !*, !^, var, val : are functions

wrappers for Expr constructors that

perform additional simplification

Minimal complete definition

eval, simplify, partDiff, simplified

Methods

eval :: Map String a -> Expr a -> a

simplify :: Map String a -> Expr a -> Expr a

partDiff :: String -> Expr a -> Expr a

(!+) :: Expr a -> Expr a -> Expr a

(!*) :: Expr a -> Expr a -> Expr a

(!^) :: Expr a -> Expr a -> Expr a

val :: a -> Expr a

var :: String -> Expr a

simplified :: Map String a -> Expr a -> Bool

Checks if the expression is already in simplified form