ExprDiff

Description

This module contains the type class DiffExpr. DiffExpr has methods over the Expr datatype that assist with the construction and evaluation of differentiable expressions.

Documentation

class DiffExpr a where

Minimal complete definition

eval, simplify, partDiff

Methods

eval :: Map String a -> Expr a -> a

Takes a dictionary of variable identifiers and values, and uses it to compute the Expr fully.

simplify :: Map String a -> Expr a -> Expr a

Takes a possibly incomplete dictionary and uses it to reduce Expr as much as possible.

partDiff :: String -> Expr a -> Expr a

Given an identifier, differentiates IN TERMS of that identifier.

(!+) :: Expr a -> Expr a -> Expr a

Addition wrapper with simplification

(!-) :: Expr a -> Expr a -> Expr a

Subtraction wrapper with simplifition

(!*) :: Expr a -> Expr a -> Expr a

Multiplication wrapper with simplifition

(!/) :: Expr a -> Expr a -> Expr a

Division wrapper with simplifition

val :: a -> Expr a

Value wrapper

var :: String -> Expr a

Variable wrapper

mySin :: Expr a -> Expr a

Sine wrapper with simplification

myCos :: Expr a -> Expr a

Cosine wrapper with simplification

myExp :: Expr a -> Expr a

Natural exponentiation wrapper with simplification

myLn :: Expr a -> Expr a

Natural logarithm wrapper with simplification