ExprDiff

Description

This module is geared towards calculus. The differential expression which include partDiff and deriv, which contain type classes and instances. There also also anti derivative present intr, to aide in surface area computation there is areay and areax which has a type class and instances to perform calculus calculations. There is basic functions for computiations in type class and instances !+, !*, !-, !^, tinaCos, tinaSin, tinaAcos, tinaAsin, tinaLn, neg. These are all simplified and evaluated with the eval and simplify classes. My special features that i included are my !-; subtraction, !^; a^ b, tinaCos; cosine, tinaSin; sine, tinaAcos; acos, tinaAsin; acos, tinaLn; log , neg; negates. Also included usefull tools such as intr; antiderivative, deriv; derivatives, areay; surface area of revolution on y, and areax ; surface area of revolution on x.

Documentation

data EvalResult a

Constructors

EvalError String
evaluates the input and returns the result

Result a
data type evaluates result as either and returns the error

class DiffExpr a where

type class of certian expressions in calculus

Minimal complete definition

eval, simplify, partDiff, deriv, intr, areay, areax

Methods

eval

Arguments

:: Map String a
evaluates input as Map.Map string a

-> Expr a
evaluates the expression

-> a
return some result of the differential equation

simplify

Arguments

:: Map String a

-> Expr a
input expression

-> Expr a
returns a simplified expression

partDiff

Arguments

:: String
input is a string of var to be evaluated

-> Expr a
input an expression to be deifferentiated, for example input simplify (Map.singleton "x" 1) (Sub (Add (Const 35) (Const 7)) (Const 5))

-> Expr a
and returns an expression which evaluates the partial differentiation of an equation given some variable

deriv

Arguments

:: Expr a
deriv inputs an expresion ex input deriv (Const 2)

-> Expr a
simply returns the an expression which represents the derivative of the expression,

intr

Arguments

:: Expr a
intr takes the integral of an expression given some expression

-> Expr a
returns the integral as an expression

areay

Arguments

:: Expr a
areay takes some expression that respresents what y is

-> Expr a
outputs an expression which is the surface area of revolution of an equation about the y axis

areax

Arguments

:: Expr a
areay takes some expression that respresents what y is

-> Expr a
outputs an expression which is the surface area of revolution of an equation about the x axis | takes first expression takes second expression returns a added expression with

(!+) :: Expr a -> Expr a -> Expr a

(!-) :: Expr a -> Expr a -> Expr a

takes first expression takes second expression returns a subtracted expression with !-

(!*) :: Expr a -> Expr a -> Expr a

takes first expression takes second expression returns a multiplied expression with !*

(!^) :: Expr a -> Expr a -> Expr a

takes first expression takes second expression returns a multiplied expression with !* ex a ^ b

tinaCos

Arguments

:: Expr a
cosine function expression

-> Expr a
returns a cosine b evaluated expression

Cos x used to evaluate cosine from Map.fromList and simplifies it

tinaAcos

Arguments

:: Expr a
acos function expression

-> Expr a
returns a acos b evaluated expression

Acos x used to evaluate acos from Map.fromList and simplifies it

tinaAsin

Arguments

:: Expr a
asin function expression input

-> Expr a
returns a asin b evaluated expression

Asin x used to evaluate asin from Map.fromList and simplifies it

tinaSin

Arguments

:: Expr a
sin function expression input

-> Expr a
returns a sin b evaluated expression

Sin x used to evaluate sine from Map.fromList and simplifies it

tinaLn

Arguments

:: Expr a
log function expression input

-> Expr a
returns a log b evaluated expression

Ln x used to evaluate ln from Map.fromList and simplifies it *** Other

tinaE

Arguments

:: Expr a
exp function expression input

-> Expr a
returns a exp b evaluated expression

E x used to evaluate e^b from Map.fromList and simplifies it

val

Arguments

:: a
represents value input

-> Expr a
returns a expression, ex given a number

Const x used to evaluate some constant

var

Arguments

:: String
string evaluated at some value ex "x"

-> Expr a
returns it as an expression

Var x used to evaluate soem variable

neg

Arguments

:: Expr a
input is some expression

-> Expr a
returns the negated expression

Neg x used to evaluate negative expression from Map.fromList