ExprDiff

Description

differentiable expressions and the simplification of expressions. Copyright : (c) Nikhail Singh 2018 License : WTFPL Maintainer : singhn18mcmaster.ca Stability : experimental Portability : POSIX.

Documentation

class DiffExpr a where

This is the class with methods for evaluation and differentiation of an expression

Minimal complete definition

eval, simplify, partDiff

Methods

eval :: Map String a -> Expr a -> a

evaluates the expression

simplify :: Map String a -> Expr a -> Expr a

simplifies the expression

partDiff :: String -> Expr a -> Expr a

partially differentiates the expression with repsect to the variable specified by the string

(!+) :: Expr a -> Expr a -> Expr a

Combines expressions for addition

(!-) :: Expr a -> Expr a -> Expr a

Combines expressions for subtraction

(!/) :: Expr a -> Expr a -> Expr a

Combines expressions for division

(!*) :: Expr a -> Expr a -> Expr a

Combines expressions for multiplication

(!^) :: Expr a -> Expr a -> Expr a

Combines expressions for powers

val :: a -> Expr a

Generates expressions for addition

sine :: Expr a -> Expr a

creates expressions for sin

cosine :: Expr a -> Expr a

creates expressions for cos

tangent :: Expr a -> Expr a

creates expressions for tan

ln1 :: Expr a -> Expr a

creates expressions for ln

exp1 :: Expr a -> Expr a

creates expressions for ln

var :: String -> Expr a

creates expressions for variables

How to evaluate and differentiate different methods

generalizing Class

class Num a => ExprEval a where

Class for generalizing expression evaluation for all numeric types

Minimal complete definition

evalExp, evalLn, evalSin, evalCos, evalTan

Methods

evalExp :: a -> a

eval for e^x

evalLn :: a -> a

eval for Ln

evalSin :: a -> a

eval for sin

evalCos :: a -> a

eval for cos

evalTan :: a -> a

eval for tan