ExprDiff

Description

Contains a type class and instances for differentiable expressions. Within the instances are several useful expressions for the evaluation, simplification, and partial differentiation of Expr type variables. This module also contains various vector operations including evaluation/simplifcation of a vector containing expressions, simple addition and subtraction, dot product.

Documentation

class DiffExpr a where

Minimal complete definition

eval, evalVect, dotProd, crossProd, simplify, partDiff

Methods

eval

Arguments

:: Map String a
The map contianing the values of all variables used.

-> Expr a
The expression to be evaluated.

-> a
The result of the evaluation of the expression.

Evaluates non-vector expressions of type Expr a. Takes a map contianing the values of any variables used in the expression. Returns a value of type a. Note that this function will error if undefined variables exist within the given expression or if an invalid Expr is given (ie. a vector. See evalVect for these types of expressions).

evalVect

Arguments

:: Map String a
The map contianing the values of all variables used.

-> Expr a
The vector expression to be evaluated.

-> [a]
The result of the evaluation of the expression.

Evaluates a vector of type Expr a. Takes a map containing the values of any variables used within the vector. Returns a value of type [a]. Note that this function will error if there is an undefined variable within the vector or if an invalid expression is given (ie. anything not confound to vectors, vector addition, or vector subtraction. See eval for these types of expressions).

dotProd

Arguments

:: Map String a
The map contianing the values of all variables used.

-> Expr a
The first vector in the dot product.

-> Expr a
The second vector in the dot product.

-> a
The result of the dot product.

Evaluates the dot product of two given vectors (with the types Expr a). Returns a value of type a. Takes a map containing the values of any variables used within the vectors. Note that this function will error if the inputed expressions are not two vectors.

crossProd

Arguments

:: Map String a
The map contianing the values of all variables used.

-> Expr a
The first vector in the cross product.

-> Expr a
The second vector in the cross product.

-> Expr a
The result of the cross product.

Evaluates the cross prooduct of two given 3-dimensional vectors (with types Expr a). Returns a value of type Expr a. Takes a map containing the values of any variables used within the vectors. Note that this function will error if the inputed expressions are not two vectors.

simplify

Arguments

:: Map String a
The map contianing the values of all variables used.

-> Expr a
The expression to be simplified.

-> Expr a
The result of the simplification of the expression.

Simplifies expressions of type Expr a. Takes a map contianing the values of any variables used in the expression. If a variable in the expression is not defined in the map, it will be left in it's String form. Returns a simplified Expr a. Note that this function will not work on vectors (see evalVect for all your vector needs).

partDiff

Arguments

:: String
The variable to be differentiated with respect to.

-> Expr a
The expression to be differentiated.

-> Expr a
The resulting partially differentiated expression.

Partially differentiates an expression of type Expr a with respect to a given variable. Takes a String with the variable to be differentiated. Returns the differentiated expression of type Expr a.

(!+)

Arguments

:: Expr a
The first expression in the addition.

-> Expr a
The second expression in the addition.

-> Expr a
The simplified addition of both expressions.

A prettier form of the Add constructor found in the Expr type. Takes two expressions of type Expr a, wraps them in the Add constructor and simplifies the new expression.

(!-)

Arguments

:: Expr a
The first expression in the subtraction.

-> Expr a
The second expression in the subtraction.

-> Expr a
The simplified subtraction of both expressions.

A prettier form of the Sub constructor found in the Expr type. Takes two expressions of type Expr a, wraps them in the Sub constructor and simplifies the new expression.

(!*)

Arguments

:: Expr a
The first expression in the multiplication.

-> Expr a
The second expression in the multiplication.

-> Expr a
The simplified multiplication of both expressions.

A prettier form of the Mult constructor found in the Expr type. Takes two expressions of type Expr a, wraps them in the Mult constructor and simplifies the new expression.

(!/)

Arguments

:: Expr a
The first expression in the division.

-> Expr a
The second expression in the division.

-> Expr a
The simplified division of both expressions.

A prettier form of the Div constructor found in the Expr type. Takes two expressions of type Expr a, wraps them in the Div constructor and simplifies the new expression.

(!^)

Arguments

:: Expr a
The base of the power.

-> Expr a
The exponent of the power.

-> Expr a
The simplified division of both expressions.

A prettier form of the Pow constructor found in the Expr type. Takes two expressions of type Expr a, wraps them in the Pow constructor and simplifies the new expression.

cos'

Arguments

:: Expr a
The expression to be evaluated through the Cos function.

-> Expr a
The given expression wrapped wihtin the Cos constructor

A prettier form of the Cos constructor found in the Expr type. Takes an expression and wraps it in the Cos constructor.

sin'

Arguments

:: Expr a
The expression to be evaluated through the Sin function.

-> Expr a
The given expression wrapped wihtin the Sin constructor

A prettier form of the Sin constructor found in the Expr type. Takes an expression and wraps it in the Sin constructor.

log'

Arguments

:: Expr a
The expression to be evaluated through the Log function.

-> Expr a
The given expression wrapped wihtin the Log constructor

A prettier form of the Log constructor found in the Expr type. Takes an expression and wraps it in the Log constructor.

exp'

Arguments

:: Expr a
The expression to be evaluated through the Exp function.

-> Expr a
The given expression wrapped wihtin the Exp constructor

A prettier form of the Exp constructor found in the Expr type. Takes an expression and wraps it in the Exp constructor.

val

Arguments

:: a
The value to be treated as a constant.

-> Expr a
The value wrapped in the Const constructor.

A prettier form of the Val constructor found in the Expr type. Takes a value of type a and wraps it in the Const constructor returning an expression of type Expr a.

var

Arguments

:: String
The variable name.

-> Expr a
The variable wrapped in the Var constructor.

A prettier form of the Val constructor found in the Expr type. Takes a variable of type String and wraps it in the Var constructor returning an expression of type Expr a.

vect

Arguments

:: [Expr a]
The list of components for the vector.

-> Expr a
The component list wrapped in the Vect constructor.

A prettier form of the Vect constructor found in the Expr type. Takes a list of type [Expr a] and returns a vector in the Expr type using the given list as it's components.