ExprDiff

Description

Documentation

list :: Map [Char] Double

Class Declaration For DiffExpr

class DiffExpr a where

Minimal complete definition

eval, simplify, partDiff

Methods

eval

Arguments

:: Map String a
Takes a Dictionary With Strings and Numbers

-> Expr a
And Takes an Expression

-> a
Returns a Value in The Type Used In The Inputs

Function For Evaluating an Expression Given a Dictionary

simplify

Arguments

:: Map String a
Takes a Dictionary With Strings and Numbers (Possibly Empty)

-> Expr a
And Takes an Expression

-> Expr a
Returns a Possibly Simplified Expression

Function For Simplifying an Expression Given a Dictionary That May be Empty

partDiff

Arguments

:: String
Takes a Variable To be Differentiated as a String

-> Expr a
And Takes an Expression

-> Expr a
Returns a Partial Differentiated Expression in Terms of The String Provided

Function For Calculating The Partial Differentiation of an Expression

(!+) :: Expr a -> Expr a -> Expr a

Generate Add Type For an Expression

(!*) :: Expr a -> Expr a -> Expr a

Generate Mult Type For an Expression

(!^) :: Expr a -> Expr a -> Expr a

Generate Raise Type For an Expression

val :: a -> Expr a

Function For Turning a Value Into a Const Type for an Expresion

var :: String -> Expr a

Function For Turning a String Into a Var Type for an Expresion

sine :: Expr a -> Expr a

Function For Turning an Expr Into a Sin Type for an Expresion

cosine :: Expr a -> Expr a

Function For Turning an Expr Into a Cos Type for an Expresion

natlog :: Expr a -> Expr a

Function For Turning an Expr Into a Natlog Type for an Expresion

expnat :: Expr a -> Expr a

Function For Turning an Expr Into a Exp Type for an Expresion