ExprDiff

Description

Documentation

class DiffExpr a where

This class operates over the Expr Data type, contains definitions for method for the Expr

Minimal complete definition

eval, simplify, partDiff

Methods

eval

Arguments

:: Map String a
Dictionary of indentifiers

-> Expr a
Expression to be evaluated

-> a
resulting value

Given a dictionary of variable indentifiers and an Expr, evaluate that Expr to a single value

simplify

Arguments

:: Map String a
Dictionary of indentifiers

-> Expr a
Expression to be simplifyed

-> Expr a
resulting Expression

Given a posibly incomplete dictionary of varuable indetifiers and Expr, evluate the Expr as much as possible and reduce

partDiff

Arguments

:: String
String with the indentifier

-> Expr a
Expression to be partially differentiated

-> Expr a
Resulting Expression

Takes a string of a sngle variable indentifier and an Expr and differentiats that Expr in terms of the indentifier

(!+)

Arguments

:: Expr a
First Expression

-> Expr a
Second Expression

-> Expr a
Resulting simplified Add Expression

corresponds to the Add type wrapper to optimize Add type. tries to simplify the Expr

(!-)

Arguments

:: Expr a
First Expression given

-> Expr a
Second Expression given

-> Expr a
Resulting simplified Sub Expression

corresponds to the Sub type wrappers to optimize Sub type. tries to simplify the Expr

(!*)

Arguments

:: Expr a
First Expression given

-> Expr a
Second Expression given

-> Expr a
Resulting simplified Mult Expression

corresponds to the Mult type wrappers to optimize Mult type. tries to simplify the Expr

(!/)

Arguments

:: Expr a
First Expression given

-> Expr a
Second Expression given

-> Expr a
Resulting simplified Div Expression

corresponds to the Div type wrappers to optimize Div type. tries to simplify the Expr

(!^)

Arguments

:: Expr a
First Expression given

-> Expr a
Second Expression given

-> Expr a
Resulting simplified Exp Expression

corresponds to the Exp type wrappers to optimize Exp type. tries to simplify the Expr

mycos

Arguments

:: Expr a
Expression given

-> Expr a
Resulting simplified Cos Expression

corresponds to the Cos type wrappers to optimize Cos type. tries to simplify the Expr

mysin

Arguments

:: Expr a
Expression given

-> Expr a
Resulting simplified Sin Expression

corresponds to the Sin type wrappers to optimize Sin type. tries to simplify the Expr

mylog

Arguments

:: Expr a
Expression given

-> Expr a
Resulting simplified Log Expression

corresponds to the Log type wrappers to optimize Log type. tries to simplify the Expr

natexp

Arguments

:: Expr a
Expression given

-> Expr a
Resulting simplified NatExp Expression

corresponds to the NatExp type wrappers to optimize Natexp type. tries to simplify the Expr

matrix

Arguments

:: [[a]]
Given matrix represented by a list of lists

-> Expr a
Resulting Matrix Expression

Takes a List of Lists and wraps it in a Matrix constructor

val

Arguments

:: a
given value

-> Expr a
Resulting Expression

Takes a value of type a and wraps it in Const constructor

var

Arguments

:: String
Giving String

-> Expr a
Resulting Expression

Takes a string and wraps it in a Var constructor

Miscellaneous Functions

listadd :: Num a => [a] -> [a] -> [(a, a)]

This fuction takes 2 lists and combines then to form a list of tuples with matching element indexs for each tuple. Need this for addmatrix. e.g listadd [1,1] [2,2] == [(1,2), (1,2)].

listcombine2 :: Num a => [(a, a)] -> [a]

This fuction adds the elements in the tuples to form a list. e.g listcombine2 [(1,2), (3,4)] == [3,7].

listsubtract :: Num a => [(a, a)] -> [a]

This function is the same as listadd only for subtraction. e.g listsubract [(1,1), (3,6)] == [0,-3].

addmatrix :: Num a => [[a]] -> [[a]] -> [[a]]

This function takes a matrix which is represented by a list of list, with each list being a row, and adds them together to get a martix (List of Lists). e.g addmatrix [[1,2],[3,4] [[1,1],[1,1]] == [[2,3],[4,5]].

submatrix :: Num a => [[a]] -> [[a]] -> [[a]]

This function takes a matrix which is represented by a list of list, with each list being a row, and subtracts them to get a matrix (List of Lists). e.g submatrix [[3,4]] [[1,2]] == [[2,2]]

multMatrixByC :: Num a => a -> [[a]] -> [[a]]

This function takes a constant of Num class and multiplies a Matrix (List of Lists) by it. e.g multMatrixByC 2 [[1,2],[3,4]] == [[2,4],[6,8]].

tupleadd :: Num a => (a, a) -> a

This function takes a Num class tuple and adds the elements together. e.g tupleadd (5,4) == 9

tuplesub :: Num a => (a, a) -> a

This function takes a Num class tuple and subtracts the elements. e.g tuplesub (4,2) == 2

matrixMult :: Num a => [[a]] -> [[a]] -> [[a]]

This function takes 2 Matrices (List of Lists) and, using matrix multiplication, multiplies them together e.g matrixMult [[1,2],[3,4]] [[2,2], [2,3]] == [[6,8],[14,18]].

matstep2 :: Num a => [a] -> [[a]] -> [a]

This function helped me in writing matrixMult. Do not use it individually.

matstep1 :: Num a => [a] -> [a] -> a

This function helped me in writing matrixMult. Do not use it individually.