ExprParser

Description

parses using parseExprD and parseExprF are parsers used to parse a string into float or double. This parser has the ability to search for key words ex Const Var .. and convert it into an expression which can be evaluated

Top Parsers for Double and Float

parseExprD

Arguments

:: String
The string to be parsed

-> Expr Double
The resulting parsed Double Expr

parseExprF

Arguments

:: String
The string to be parsed

-> Expr Float
The resulting parsed Float Expr

Lower Parsers for float and double

dFactor

Arguments

:: Parser (Expr Double)
Parse expression as a Double

dConst :: ParsecT String () Identity (Expr Double)

fFactor

Arguments

:: Parser (Expr Float)
Parse expression as a Float

fConst :: ParsecT String () Identity (Expr Float)

var

Arguments

:: Parser (Expr a)
parses expression for Var ss

iden

Arguments

:: Parser String
parses string for alphaNum

setexpr

Arguments

:: Parser (Expr a)
parses expression as an input

-> Parser (Expr a)

allfunc

Arguments

:: Parser (Expr a -> Expr a -> Expr a)
parses for functions such as !- !+ !^ !*

trigOp :: String -> (Expr a -> Expr a) -> Parser (Expr a) -> Parser (Expr a)

evaluates for the number involved in trig expression, input is a string of the name of the trig ident and value to evaluate, expr a to expr a, parses expr a, return parser expr a

alltrig

Arguments

:: Parser (Expr a)
input is a parser of a cetrian expression

-> Parser (Expr a)
output a parser expression, aim is to match any of the trig expressions with the input

mycos

Arguments

:: Parser (Expr a)
input is a parser expression

-> Parser (Expr a)
output is also a parser expression, aim is to chack if "cos" is in the string input originally

mysin

Arguments

:: Parser (Expr a)
input is a parser expression

-> Parser (Expr a)
output is also a parser expression, aim is to chack if "sin" is in the string input originally

myacos

Arguments

:: Parser (Expr a)
input is a parser expression

-> Parser (Expr a)
output is also a parser expression, aim is to chack if "acos" is in the string input originally

myasin

Arguments

:: Parser (Expr a)
input is a parser expression

-> Parser (Expr a)
output is also a parser expression, aim is to chack if "asin" is in the string input originally

myln

Arguments

:: Parser (Expr a)
input is a parser expression

-> Parser (Expr a)
output is also a parser expression, aim is to chack if "ln" is in the string input originally

mye

Arguments

:: Parser (Expr a)
input is a parser expression

-> Parser (Expr a)
output is also a parser expression, aim is to chack if "tinaE" is in the string input originally

parens

Arguments

:: Parser a
input is a parser a

-> Parser a

symbol

Arguments

:: String
input is a string

-> Parser String
output is a parser string, aim is to parse out the spaces in the given string for an input

Utility Parsers

dec

Arguments

:: Parser String
this ia a parcer string that is used to parce decimals using the "." (found similar in Curtis' code)

fore

Arguments

:: Parser String
parses string for e by trying to see if its a digits or negDigit, and checking if theres a char e

nil

Arguments

:: Parser String
checks if theres nothing in the string

decimal

Arguments

:: Parser String
parses string for decimal using dec , nill and fore

negDecimal :: Parser String

float

Arguments

:: Parser Float
parses string for float

double

Arguments

:: Parser Double
parses string for double

digits

Arguments

:: Parser String
parses string for digits

negDigits

Arguments

:: Parser String
parses strign for negative digits using "-" symbol