ExprParser

Description

Strings to be parsed must be formated in a specific way. Below the parsing functions is a description of the format a string needs to be in to parse into the Expr type.

Documentation

parseExprD

Arguments

:: String
The string to be parsed.

-> Expr Double
The resulting Expr Double

Formats a given string into an expression of type Expr Double. The format of the string must be:

Addition -> Signified by !+

Subtraction -> Signified by !-

Multiplication -> Signified by !*

Division -> Signified by !/

Power -> Signified by !^

Cos -> Signified by cos' (expression)

Sin -> Signified by sin' (expression)

Log -> Signified by log' (expression)

Exp -> Signified by exp' (expression)

Vectors -> Signified by vect [list of expressions]

Constants -> Signified by simplly writing the constant value.

Variable -> Signified by var (variable name as string)

parseExprF

Arguments

:: String
The string to be parsed.

-> Expr Float
The resulting Expr Float

Formats a given string into an expression of type Expr Float. The format of the string must be:

Addition -> Signified by !+

Subtraction -> Signified by !-

Multiplication -> Signified by !*

Division -> Signified by !/

Power -> Signified by !^

Cos -> Signified by cos' (expression)

Sin -> Signified by sin' (expression)

Log -> Signified by log' (expression)

Exp -> Signified by exp' (expression)

Vectors -> Signified by vect [list of expressions]

Constants -> Signified by simply writing the constant value.

Variable -> Signified by var (variable name as string)