ExprParser

Description

Main Parsing Functions

parseExprD :: String -> Expr Double

Parses a String Into an Expr Double Type Using The exprD Parser

parseExprF :: String -> Expr Float

Parses a String Into an Expr Float Type Using The exprF Parser

parseExprI :: String -> Expr Integer

Parses a String Into an Expr Integer Type Using The exprI Parser

Main Parsers Used in The Parsing Functions

exprD :: Parser (Expr Double)

Parses All Valid Strings Into Expr Double Types Using Text.Parsec.Expr Parser Maker

table :: [[Operator String u Identity (Expr a)]]

factord :: ParsecT String () Identity (Expr Double)

Statement For Parsing String With Expr Double Type to Actual Type

exprF :: Parser (Expr Float)

Parses All Valid Strings Into Expr Types Using Text.Parsec.Expr Parser Maker

factorf :: ParsecT String () Identity (Expr Float)

Statement For Parsing String With Expr Float Type to Actual Type

exprI :: Parser (Expr Integer)

Parses All Valid Strings Into Expr Types Using Text.Parsec.Expr Parser Maker

factori :: ParsecT String () Identity (Expr Integer)

Parsers For Extracting Various Num Types and Vars (Parsers Used In exprD, exprF, exprI)

numd :: Parser (Expr Double)

Parses All Valid Strings Into a Double

numf :: Parser (Expr Float)

Parses All Valid Strings Into a Float

numi :: Parser (Expr Integer)

Parses All Valid Strings Into an Integer

varOpD :: Parser (Expr Double)

Parses All Valid Strings to a Variable Identifier For Expr Double Types

varOpF :: Parser (Expr Float)

Parses All Valid Strings to a Variable Identifier For Expr Float Types

varOpI :: Parser (Expr Integer)

Parses All Valid Strings to a Variable Identifier For Expr Integer Types

Smaller Accessory Parsers Used in The Larger Parsers

double :: Parser String

Takes a String and Parses it Into a Format So That it Can Be Used With Read

symbol :: String -> Parser String

Takes a String With a Symbol and Parses it

digits :: Parser String

Takes a String With Digit(s) and Parses it

negDigits :: Parser String

Takes a String With Negative Digit(s) and Parses it

integer1 :: Parser String

Takes a String With Integer(s) and Parses it