ExprTest

Description

Documentation

testvr :: Map [Char] Double

Test Cases

evalProp1 :: Double -> Double -> Bool

Checks if two variables can be added rightly when given their value.

testevalProp1 :: IO ()

evalProp2 :: Double -> Double -> Bool

Checks if two variables can be multiplied rightly when given their value.

testevalProp2 :: IO ()

evalProp3 :: Double -> Bool

Checks if e^x can be valued rightly.

testevalProp3 :: IO ()

evalProp4 :: Double -> Double -> Bool

Checks if e^x can be valued rightly, when redundant vaialbe value is given in the dictionary.

testevalProp4 :: IO ()

evalProp5 :: Double -> Double -> Bool

Checks if the second variable's value is right, when redundant vaialbe value is given earlier in the dictionary.

testevalProp5 :: IO ()

evalProp6 :: Double -> Double -> Bool

Checks if negative constants work alright while being evaluated.

testevalProp6 :: IO ()

evalProp7 :: Double -> Double -> Bool

Checks if cos works.

testevalProp7 :: IO ()

evalProp8 :: Double -> Double -> Bool

Checks if sin works.

testevalProp8 :: IO ()

diffProp1 :: String -> Double -> Bool

Checks if doing partial differentiation to a constant returns a zero.

testdiffProp1 :: IO ()