ExprTest

Description

Module : ExprTest

Description : Contains Some Tests to verify the credibility of the functions

Copyright : (c) Hariesh Jayanthan @2018

License : WTFPL

Maintainer : Jayanthh@mcmaster.ca

Stability : experimental

Portability : DOS

Documentation

evalFcheck

Arguments

:: Float

-> Float

-> Float
Quick Check produces 3 random Floats that are inputted into the equation, which is then compared against the normal mathematical equation to see if they produce the same answer to some tolerance

-> Bool
If they are similar enough, True is returned, else False is returned.

evalDcheck

Arguments

:: Double

-> Double

-> Double
Quick Check produces 3 random Double Values that are inputted into the equation, which is then compared against the normal mathematical equation to see if they produce the same answer to some tolerance

-> Bool
True is returned if the answers are within tolerance of each other

dxDouble

Arguments

:: Double
this time a double variable value is being generated with QuickCheck rather than constants

-> Bool
If it is within tolerance the function responds with True

dxFloat

Arguments

:: Float
a float variable value is being generated with QuickCheck rather than constants

-> Bool
If it is within tolerance the function responds with True