ExprTest

Description

Documentation

testPropMult :: Double -> Double -> Bool

Property For QuickCheck to check the Mult contructor

testPropCos :: Double -> Bool

Property For QuickCheck To check the Cos contructor

evalProp1 :: Double -> Double -> Bool

Checks the evaluation of an expression of type Expr Double

simpProp1 :: (Floating a, Show a, Ord a) => a -> a -> Bool

Sample expressions evaluated to test effectiveness of simplification

simpProp2 :: Double -> Double -> String -> Bool

Checks if a variable is not in the list of vars it should be returned as a variable (Idea from https://github.com/deleeuwj1)