ExprTest

Description

TODO This module contains tests to check the functionality of the library

Documentation

sampleExpr1 :: Expr Double

tests Add

listToExpr1 :: [Double] -> Expr Double

tests ability to take a list and convert into an Expr

evalProp1 :: Double -> Double -> Bool

tests validity of Add

testEvalProp1 :: IO ()

evalProp2 :: Double -> Double -> Bool

tests validity of Mult

testEvalProp2 :: IO ()

evalProp6 :: Double -> Bool

tests validity of Cos

testEvalProp6 :: IO ()

evalProp7 :: Double -> Bool

tests validity of Sin

testEvalProp7 :: IO ()