ExprTest

Description

Documentation

evalDProp

Arguments

:: Double

-> Double
Const Types With Random Doubles Generated By QuickCheck

-> Bool

Property For QuickCheck To Check For Evaluation of a Sample Expression of Type 'Expr Double'

evalFProp

Arguments

:: Float

-> Float
Const Types With Random Floats Generated By QuickCheck

-> Bool

Property For QuickCheck To Check For Evaluation of a Sample Expression of Type Expr Float

pdDProp

Arguments

:: Double
Dict With Random Double Generated For Value of x

-> Bool

Property For QuickCheck To Check For Evaluation of a Partially Differentiated Sample Expression of Type Expr Double

pdFProp

Arguments

:: Float
Dict With Random Float Generated For Value of x

-> Bool

Property For QuickCheck To Check For Evaluation of a Partially Differentiated Sample Expression of Type Expr Float

parseDProp

Arguments

:: Double

-> Double

-> Double
String With Random Doubles Generated By QuickCheck

-> Bool

Property For QuickCheck To Check For The Parsing of a String (With Doubles) in The Format Specified in ExprParser

parseFProp

Arguments

:: Float

-> Float

-> Float
String With Random Doubles Generated By QuickCheck

-> Bool

Property For QuickCheck To Check For The Parsing of a String (With Floats) in The Format Specified in ExprParser

simpDProp

Arguments

:: Double

-> Double
Const Types With Random Doubles Generated By QuickCheck

-> Bool

Property For QuickCheck To Check if The Evaluation of a Simplification (Expr Double) Gives The Right Result

simpFProp

Arguments

:: Float

-> Float
Const Types With Random Floats Generated By QuickCheck

-> Bool

Property For QuickCheck To Check if The Evaluation of a Simplification (Expr Float) Gives The Right Result