ExprType

Description

DataType Declaration

data Expr a

Constructors

Add (Expr a) (Expr a)
Binary Addition

Mult (Expr a) (Expr a)
Binary Multiplication

Const a
Value Wrapper

Var String
Variable Wrapper

Cos (Expr a)
Cosine operation

Sin (Expr a)
Sine operation

Log (Expr a)
Ln operation

Exp (Expr a)
e^x

Pow (Expr a) (Expr a)
An expression to power of another expression

Utility Functions

getVars :: Expr a -> [String]

Grab the variables out from an Expr, make them a list

(!+) :: Num a => Expr a -> Expr a -> Expr a

An function used to add two expressions into a new expression. Only used in test cases.

(!*) :: Num a => Expr a -> Expr a -> Expr a

An function used to multiply two expressions into a new expression. Only used in test cases.

(!-) :: Num a => Expr a -> Expr a -> Expr a

An function used to subtract one expression from another, retuns a new expression. Only used in test cases.

(!/) :: Num a => Expr a -> Expr a -> Expr a

An function used to devide one expression from another, retuns a new expression. Only used in test cases.

isThere :: Map String a -> String -> Bool

An auxiliary function used to determine if a giving string is in the given variable list. Returns a Bool.

areThere :: Map String a -> [String] -> Bool

An auxiliary function used to determine if ALL given strings are in the given variable list. Returns a Bool.