ExprType

Description

Documentation

data Expr a

Datatype for regularly used expressions

Constructors

Add (Expr a) (Expr a)
simple addition with 2 expr types

Mult (Expr a) (Expr a)
simple Mulitplication with 2 expr types

Const a
to wrap a constant

Var String
identifies a variable

Cos (Expr a)
Constructor for cos

Sin (Expr a)
Constructor

Ln (Expr a)
natural log

Exp (Expr a)
e (to the power of) (Expr)

Exponent (Expr a) (Expr a)

getVars :: Expr a -> [String]

getVars Takes the variables from an expression recursively if the expression is in Var constructor form it returns the variable, otherwise it strips it down until it is in Var form if the expression is a constant, there is no variable, so it returns nothing.