ExprType

Description

differentiable expressions Copyright : (c) Yifan Jiang 2018 License : WTFPL Maintainer : jiany25mcmaster.ca Stability : experimental Portability : POSIX TODO write a longer description of the module, containing some commentary with some markup.

DataType Decleration

data Expr a

A datatype for common numeric expression

Constructors

Const a
Value wrapper: Const Num a => a such as Const Integer Const Float, etc.

Var String
Variable Identifier: ex Var "x" Var "y", etc.

Add (Expr a) (Expr a)
Standard binary addition

Mult (Expr a) (Expr a)
Standard binary multiplication

Div (Expr a) (Expr a)
Standard binary division

Sin (Expr a) (Expr a)
Sine function: f(x)=sinx

Cos (Expr a) (Expr a)
Cosine function: f(x)=cosx

Log (Expr a) (Expr a)
General logarithmic function: f(x)=loga(x)

Ln (Expr a) (Expr a)
Natural logarithmic function: f(x)=ln(x)(where a = e)

Exp (Expr a) (Expr a)
General exponential function: f(x)=a^x

NatExp (Expr a) (Expr a)
Natural expoential function: f(x)=e^x(where a = e)

Miscellaneous Functions

getVars :: Expr a -> [String]

A function that obtains the varibale identifiers and forms a list of strings.