ExprType

Description

This module contains the datatype Expr, which wraps different operations in an expression tree. It encodes addition, subtraction, multiplication, division, sine, cosine, natural exponentiation, the natural logarithm, constants, and variables.

DataType Declaration

data Expr a

A datatype for common numeric expression

Constructors

Add (Expr a) (Expr a)
Binary Addition

Subt (Expr a) (Expr a)
Binary Subtraction

Mult (Expr a) (Expr a)
Binary Multiplication

Div (Expr a) (Expr a)
Division

Sin (Expr a)
Sine Function

Cos (Expr a)
Cosine Function

Exp (Expr a)
Natural Exponentiation: Exp a = e^a

Ln (Expr a)
Natural Logarithm

Const a
Value Wrapper

Var String
Variable Identifier

Miscellaneous Functions

getVars :: Expr a -> [String]

A function that retrieves variable identifiers from an Expr.