ExprType

Description

Contains the Expr data type as well as useful functions relating to basic use of Expr's.

Documentation

data Expr a

A datatype for common numeric expression

Constructors

Add (Expr a) (Expr a)
Binary Addition

Sub (Expr a) (Expr a)
Binary Subtraction

Mult (Expr a) (Expr a)
Binary Multiplication

Div (Expr a) (Expr a)
Binary Division

Cos (Expr a)
Triganometric Cosine Function

Sin (Expr a)
Triganometric Sine Function

Log (Expr a)
Logorithmic Function

Exp (Expr a)
Exp (Natural) Function

Vect [Expr a]
Vector Wrapper

Pow (Expr a) (Expr a)
Power Operator

Const a
Value Wrapper

Var String
Variable Identifier

getVars

Arguments

:: Expr a
The expression to be searched for variables.

-> [String]
A list of the variables found with the [String] type.

A function that takes a variable of type (Expr a) and returns a list of strings containing the variable names wrapped by Var in the expression.